Extract/move function & remove console log used for debugging
