+++ src/libs/events.js
@@ -1,3 +1,4 @@
+import _ from 'lodash/fp'
 import { useEffect } from 'react'
 import { Ajax } from 'src/libs/ajax'
 import { useRoute } from 'src/libs/nav'
@@ -62,6 +63,50 @@ export const PageViewReporter = () => {
   }, [name, params])
 
   return null
+}
+
+export const captureAppcuesEvent = (eventName, event) => {
+  // Only record "public-facing events" (and related properties) as documented by Appcues: https://docs.appcues.com/article/301-client-side-events-reference
+  const publicEvents = [
+    'flow_started',
+    'flow_completed',
+    'flow_skipped',
+    'flow_aborted',
+    'step_started',
+    'step_completed',
+    'step_skipped',
+    'step_aborted',
+    'step_interacted',
+    'form_submitted',
+    'form_field_submitted'
+  ]
+  if (_.includes(eventName, publicEvents)) {
+    const eventProps = {
+      'appcues.flowId': event.flowId,
+      'appcues.flowName': event.flowName,
+      'appcues.flowType': event.flowType,
+      'appcues.flowVersion': event.flowVersion,
+      'appcues.id': event.id,
+      'appcues.interaction.category': event.interaction?.category,
+      'appcues.interaction.destination': event.interaction?.destination,
+      'appcues.interaction.element': event.interaction?.element,
+      'appcues.interaction.fields': event.interaction?.fields ? JSON.stringify(event.interaction.fields) : undefined,
+      'appcues.interaction.formId': event.interaction?.formId,
+      'appcues.interaction.text': event.interaction?.text, // not documented by Appcues, but observed and useful
+      'appcues.interactionType': event.interactionType,
+      'appcues.localeId': event.localeId,
+      'appcues.localeName': event.localeName,
+      'appcues.name': event.name,
+      'appcues.sessionId': event.sessionId,
+      'appcues.stepChildId': event.stepChildId,
+      'appcues.stepChildNumber': event.stepChildNumber,
+      'appcues.stepId': event.stepId,
+      'appcues.stepNumber': event.stepNumber,
+      'appcues.stepType': event.stepType,
+      'appcues.timestamp': event.timestamp
+    }
+    return Ajax().Metrics.captureEvent(eventsList.appcuesEvent, eventProps)
+  }
 }
 
 export default eventsList
